Unified "ohlc" chart with an option to display candlestick or ohlc bar
Settings button activates a menu of options.
The option will be saved as part of the chart settings via the save/restore feature.
List item font
