+++ packages/perspective-viewer-d3fc/src/less/chart.less
@@ -234,6 +234,70 @@
         }
     }
 
+    .settings-container {
+        position: absolute;
+        top: 45px;
+        left: 100px;
+
+        & button {
+            background: transparent;
+            border: 1px solid transparent;
+            padding: 2px 5px;
+            font-size: 1.2em;
+            font-weight: bold;
+            color: #888;
+
+            &:hover {
+                border: 1px solid rgb(204, 204, 204);
+                color: black;
+            }
+        }
+    }
+
+    .settings-overlay {
+        position: absolute;
+        top: 0;
+        left: 0;
+        right: 0;
+        bottom: 0;
+
+        & ul {
+            position: absolute;
+            top: 85px;
+            left: 110px;
+            list-style: none;
+            padding: 0;
+            margin: 0;
+            border: 1px solid rgb(204, 204, 204);
+            background: rgb(230, 230, 230);
+        }
+
+        & li {
+            cursor: pointer;
+            padding: 5px;
+            font-family: Arial;
+            &:hover {
+                background: rgb(204, 204, 204);
+            }
+        }
+
+        & .checked {
+            display: inline-block;
+            width: 20px;
+            height: 20px;
+            vertical-align: middle;
+            
+            &.selected::before {
+                content: "\2714";
+            }
+        }
+
+        & .label {
+            height: 20px;
+            vertical-align: middle;
+        }
+    }
+
     .d3_heatmap .legend-container.legend-color {
         height: 200px;
     }
@@ -246,16 +310,19 @@
         text-align: center;
 
         & button {
-            -webkit-appearance: none;
             background: rgb(247, 247, 247);
             border: 1px solid rgb(204, 204, 204);
-            padding: 10px;
-            opacity: 0.5;
-            cursor: pointer;
+            padding: 5px 10px;
+        }
+    }
 
-            &:hover {
-                background: rgb(230, 230, 230);
-            }
+    .settings-container button, .zoom-controls button {
+        -webkit-appearance: none;
+        opacity: 0.5;
+        cursor: pointer;
+
+        &:hover {
+            background: rgb(230, 230, 230);
         }
     }
